+++ Pipaslot.Mediator/Middlewares/IExecutionMiddleware.cs
@@ -1,11 +1,9 @@
 ﻿namespace Pipaslot.Mediator.Middlewares
 {
     /// <summary>
-    /// Marks middleare as final/last wchich executes handlers. Pipeline ends with this milleware evenf it some next middlewares are registered.
-    /// This interface was introduced to connect pipeline definitions and query handler existence check.
+    /// Default interface marking middleware as last middleware in pipeline executing handlers or other operations with actions
     /// </summary>
     public interface IExecutionMiddleware : IMediatorMiddleware
     {
-        bool ExecuteMultipleHandlers { get; }
     }
 }