+++ Pipaslot.Mediator/Services/ServiceProviderExtensions.cs
@@ -13,44 +13,75 @@ namespace Pipaslot.Mediator.Services
         /// <summary>
         /// Get all registered handlers from service provider
         /// </summary>
-        public static object[] GetMessageHandlers(this IServiceProvider serviceProvider, Type? messageType)
+        public static object[] GetMessageHandlers(this IServiceProvider serviceProvider, Type? messageType, ActionToHandlerBindingType bindingType)
         {
             if (messageType == null)
             {
                 return new object[0];
             }
-            var handlerType = typeof(IMediatorHandler<>).MakeGenericType(messageType);
-            return serviceProvider.GetServices(handlerType)
-                .Where(h => h != null)
-                // ReSharper disable once RedundantEnumerableCastCall
-                .Cast<object>()
-                .ToArray();
+            var mediatorActionType = typeof(IMediatorAction);
+            var actionTypes = bindingType == ActionToHandlerBindingType.Class
+                ? new[] { messageType }
+                : bindingType == ActionToHandlerBindingType.Interface
+                    ? messageType
+                        .GetInterfaces()
+                        .Where(i => mediatorActionType.IsAssignableFrom(i))
+                        .ToArray()
+                    : throw new NotImplementedException($"Unknown binding type {bindingType}");
+
+            var handlers = new List<object>();
+            foreach (var actionType in actionTypes)
+            {
+                var handlerType = typeof(IMediatorHandler<>).MakeGenericType(actionType);
+                handlers.AddRange(serviceProvider.GetServices(handlerType)
+                    .Where(h => h != null)
+                    // ReSharper disable once RedundantEnumerableCastCall
+                    .Cast<object>()
+                    .ToArray());
+            }
+            return handlers.ToArray();
         }
 
         /// <summary>
         /// Get all registered handlers from service provider
         /// </summary>
-        public static object[] GetRequestHandlers<TResponse>(this IServiceProvider serviceProvider, Type? requestType)
-        {
-            return serviceProvider.GetRequestHandlers(requestType, typeof(TResponse));
-        }
-
-
-        /// <summary>
-        /// Get all registered handlers from service provider
-        /// </summary>
-        public static object[] GetRequestHandlers(this IServiceProvider serviceProvider, Type? requestType, Type? responseType)
+        public static object[] GetRequestHandlers(this IServiceProvider serviceProvider, Type? requestType, Type? responseType, ActionToHandlerBindingType bindingType)
         {
             if (requestType == null || responseType == null)
             {
                 return new object[0];
             }
-            var handlerType = typeof(IMediatorHandler<,>).MakeGenericType(requestType, responseType);
-            return serviceProvider.GetServices(handlerType)
-                .Where(h => h != null)
-                // ReSharper disable once RedundantEnumerableCastCall
-                .Cast<object>()
-                .ToArray();
+            var mediatorHandlerType = typeof(IMediatorHandler<,>);
+            var actionTypes = bindingType == ActionToHandlerBindingType.Class
+                ? new[] { requestType }
+                : bindingType == ActionToHandlerBindingType.Interface
+                    ? requestType
+                        .GetInterfaces()
+                        .Where(i => IsRequestInterface(i, responseType))
+                        .ToArray()
+                    : throw new NotImplementedException($"Unknown binding type {bindingType}");
+
+            var handlers = new List<object>();
+            foreach (var actionType in actionTypes)
+            {
+                var handlerType = mediatorHandlerType.MakeGenericType(actionType, responseType);
+                handlers.AddRange(serviceProvider.GetServices(handlerType)
+                    .Where(h => h != null)
+                    // ReSharper disable once RedundantEnumerableCastCall
+                    .Cast<object>()
+                    .ToArray());
+            }
+            return handlers.ToArray();
+        }
+
+        private static bool IsRequestInterface(Type type, Type responseType)
+        {
+            if (type.IsInterface)
+            {
+                var mediatorActionType = typeof(IMediatorAction<>).MakeGenericType(responseType);
+                return mediatorActionType.IsAssignableFrom(type);
+            }
+            return false;
         }
 
         public static IExecutionMiddleware GetExecutiveMiddleware(this IServiceProvider serviceProvider, Type requestType)